write function if else in file_manip

+++ movie.h
@@ -26,7 +26,7 @@ struct MovieNode {
 struct Movie search_using_name(char *str);
 struct Movie search_using_id(int search_id);
 struct MovieNode* read_file_till_num(int n);
-void write_file(char* movieName, int rating);
+void write_file(char* movieName, float rating); // rating is of float type
 struct MovieNode* read_file();
 int num_of_data();
